refactor(security): consolidate injection patterns and use shared homoglyph/safeEqual
- Create shared-patterns.ts as single source of truth for prompt injection detection
- Replace duplicated patterns in fast-path.ts, external-content.ts, and skill-scanner.ts
- Replace manual zero-width char regex in fast-path.ts with containsHomoglyphs()
- Rename DANGEROUS_PATTERNS to DANGEROUS_COMMAND_PATTERNS in permissions.ts
- admin-claim.ts already uses crypto.timingSafeEqual (no change needed)
- Strip global flag from shared patterns used in DANGEROUS_PATTERNS to avoid
  stateful .test() behavior
